(PC-17806) refactor(auth): move useDepositAmountByAge in helpers folder
